Move scheduled report to a top-level `feature`
